requirements.txt
anthropic
httpx
anyio
ddgs
python-dotenv
apscheduler
fastapi
uvicorn
tzdata
playwright
